plot over line 機能について、Hydrogenデータで正常に分布が表示できない症状を修正。CSPBVRにて実施した修正をISPBVRでも反映

+++ IS_DaemonAndSampler/InSituLib/unstruct/TFS/PlotOverLine.h
@@ -12,9 +12,12 @@
 //#include "../kvs_wrapper.h"
 namespace kvs
 {
-    using Vec3 = Vector3<float>;
-    using Vec4 = Vector4<float>;
-    using Mat3 = Matrix33<float>;
+    //using Vec3 = Vector3<float>;
+    //using Vec4 = Vector4<float>;
+    //using Mat3 = Matrix33<float>;
+    using Vec3 = Vector3<double>;
+    using Vec4 = Vector4<double>;
+    using Mat3 = Matrix33<double>;
 }
 namespace POL
 {
